+++ CountPositives.cuh
@@ -1,6 +1,6 @@
 #pragma once 
 
 
-__global__ void runPositivesKernel(int* count, const float* __restrict__ y, int N);
+__global__ void runPositivesKernel(int* count, const double* __restrict__ y, int N);
 
-int calc_positives(float* y, const SamplingRange& h_range_in);
+int calc_positives(double* y, const SamplingRange& h_range_in);
